requirements.txt
amqp==5.1.1
asgiref==3.7.2
billiard==4.1.0
black==23.9.1
celery==5.3.4
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==4.2.6
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
iniconfig==2.0.0
kombu==5.3.2
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.11.2
platformdirs==3.11.0
pluggy==1.3.0
prompt-toolkit==3.0.39
psycopg2==2.9.9
PyJWT==2.8.0
pytest==7.4.2
pytest-django==4.5.2
pytest-mock==3.11.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
redis==5.0.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
vine==5.0.0
wcwidth==0.2.8
